Fix SumologicLogging
The class for SumologicLogging was declared, but it was never included
in fastly.rb to autogenerate the methods to call anything.

I added a test, but didn't manually update the comment documentation in
fastly.rb, as that looks hand generated and out of date for arguments,
so I figured the test was more useful.